% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.R
\name{perm}
\alias{perm}
\title{Compute TFCE derived p-values using a standard permutation testing procedure.}
\usage{
perm(X, Y, extract, A, NNmatrix, nPermutations = 1000, HC4m = FALSE,
  parallel = FALSE, nCores = 1, verbOutput = 0)
}
\arguments{
\item{X}{A V-dimensional vector (V = number of vertices in the ventricular mesh) containing the values of a statistic at each vertex in the mesh.}

\item{A}{A V-dimensional vector containing the area associated with a vertex, usually its Voronoi area.}

\item{NNmatrix}{Nx2 matrix containing the mesh edges. Important: to speed up the execution please avoid repetitions like (A,B) and (B,A).}

\item{nPermutations}{number of permutations in the permutation test, default is 1000.}

\item{HC4m}{flag for triggering HC4m correction, default is FALSE.}

\item{parallel}{flag for triggering parallel computing, default is FALSE.}

\item{nCores}{flag for defining the number of cores to use, default is 1.}

\item{verbOutput}{flag for activating verbose output, default is 0 (off).}
}
\value{
If verbOutput = 0 the output is a matrix containing in its rows the pvalues computed at each vertex and the number of colums referes to the variables specified in extract. If verbOutput = 1 the output is a list where the pval field contains the the pvalues computed at each vertex, TFCEmatrix field contains a V x nPermutations matrix containing the TFCE scores computed for each permutation and the tfceScores field is a V-dimensional vector containing the TFCE scores of the non-permuted data.
}
\description{
Given a NxV imaging matrix Y (N = number of subjects, V = number of vertices in the ventricular mesh), a NxC model matrix X  (N = number of subjects, C = number of variables + intercept term) 
and the numbers of the column variables to extract, this function computes for each variable specified in extract the TFCE derived p-values map on the mesh. The output is a matrix with a number of 
columns equal to the lenght of extract and a rows equal to the number of vertices.
}
\examples{
TFCEresults = perm(X, Y, extract, A, NNmatrix, nPermutations = 1000, HC4m = FALSE)
}
\keyword{mur}
\keyword{regression}
