% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permTT.R
\name{permTT}
\alias{permTT}
\title{Compute TFCE derived p-values for t-tests using a standard permutation testing procedure.}
\usage{
permTT(X1, X2, A, NNmatrix, nPermutations = 1000, parallel = FALSE,
  nCores = 1)
}
\arguments{
\item{X1}{Input matrix 1 (ncol=# of vertices, nrow=# number of samples).}

\item{X2}{Input matrix 2 (ncol=# of vertices, nrow=# number of samples).}

\item{A}{A V-dimensional vector containing the area associated with a vertex, usually its Voronoi area.}

\item{NNmatrix}{Nx2 matrix containing the mesh edges. Important: to speed up the execution please avoid repetitions like (A,B) and (B,A).}

\item{nPermutations}{number of permutations in the permutation test, default is 1000.}

\item{parallel}{flag for triggering parallel computing, default is FALSE.}

\item{nCores}{flag for defining the number of cores to use, default is 1.}
}
\value{
An array with a number of entries equal to the number of vertices under exam.
}
\description{
Given X1, X2 matrices (ncol=# of vertices, nrow=# number of samples) this function computes at each vertex of the atlas the TFCE-derived pvalues of the t-test at each vertex between one column of X1 and one column of X2.
}
\examples{
TFCEresults = permTT(X1, X2, A, NNmatrix, nPermutations = 1000, parallel=FALSE, nCores=1)
}
\keyword{TFCE}
\keyword{t-test}
