% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFCE.R
\name{TFCE}
\alias{TFCE}
\title{Compute TFCE on a ventricular mesh.}
\usage{
TFCE(h, A, NNmatrix, E = 0.5, H = 2, deltah = 0.01)
}
\arguments{
\item{h}{A V-dimensional vector (V = number of vertices in the ventricular mesh) containing the values of a statistic at each vertex in the mesh.}

\item{A}{A V-dimensional vector containing the area associated with a vertex, usually its Voronoi area.}

\item{NNmatrix}{A Nx2 matrix containing the mesh edges. Important: to speed up the execution please avoid repetitions ofthe form (A,B) and (B,A) in the matrix.}

\item{E}{TFCE parameter, by default fixed to 0.5.}

\item{H}{TFCE parameter, by default fixed to 2.}

\item{deltah}{discrete integral sampling dimension, by default fixed to 2.}
}
\value{
A V-dimensional vector containig the TFCE score computed at each vertex.
}
\description{
Given a statistical map on a 3D mesh this function computed the related TFCE map. The 3D mesh is described as a graph using a two-columns matrix the mesh edges, each row containing the vertex ID of a edge.
}
\examples{
TFCEscores = TFCE(h, A, NNmatrix) #run with default TFCE parameters.
TFCEscores = TFCE(h, A, NNmatrix, deltah=0.02) #change discrete integral sampling.
TFCEscores = TFCE(h, A, NNmatrix, E=1, H=1, deltah=0.02) #user default TFCE parameters.
}
\keyword{TFCE}
\keyword{multiple}
\keyword{testing}
